Extract views and structs to its own file
